fix: update XRay propagator
