+++ propagator/xray/test/xray_test.rb
@@ -7,20 +7,11 @@
 require 'test_helper'
 
 describe OpenTelemetry::Propagator::XRay do
-  describe '#text_map_extractor' do
-    it 'returns an instance of TextMapExtractor' do
-      extractor = OpenTelemetry::Propagator::XRay.text_map_extractor
-      _(extractor).must_be_instance_of(
-        OpenTelemetry::Propagator::XRay::TextMapExtractor
-      )
-    end
-  end
-
-  describe '#text_map_injector, #rack_injector' do
-    it 'returns an instance of TextMapInjector' do
-      injector = OpenTelemetry::Propagator::XRay.text_map_injector
-      _(injector).must_be_instance_of(
-        OpenTelemetry::Propagator::XRay::TextMapInjector
+  describe '#text_map_propagator' do
+    it 'returns an instance of TextMapPropagator' do
+      propagator = OpenTelemetry::Propagator::XRay.text_map_propagator
+      _(propagator).must_be_instance_of(
+        OpenTelemetry::Propagator::XRay::TextMapPropagator
       )
     end
   end
